requirements.txt
torch==1.10.0
torchvision==0.11.1
numpy
pandas
scikit-learn
Pillow==8.4.0 
scikit-build 
cmake
opencv-python==4.6.0.66
tqdm 
tensorboard 
matplotlib
ml-collections
